Improved code with CLION analisys

+++ src/Splines/SplineBiCubic.hxx
@@ -71,21 +71,21 @@ namespace Splines {
     //!
     real_type
     Dx_node( integer i, integer j ) const
-    { return m_DX[size_t(this->ipos_C(i,j))]; }
+    { return m_DX[ this->ipos_C(i,j) ]; }
 
     //!
     //! Estimated `y` derivatives at node `(i,j)`
     //!
     real_type
     Dy_node( integer i, integer j ) const
-    { return m_DY[size_t(this->ipos_C(i,j))]; }
+    { return m_DY[ this->ipos_C(i,j) ]; }
 
     //!
     //! Estimated mixed `xy` derivatives at node `(i,j)`
     //!
     real_type
     Dxy_node( integer i, integer j ) const
-    { return m_DXY[size_t(this->ipos_C(i,j))]; }
+    { return m_DXY[ this->ipos_C(i,j) ]; }
 
     ///@}
